feat(command): create build command for components
