+++ src/commands/build.ts
@@ -1,7 +1,12 @@
 import * as Mustache from 'mustache';
-import writeFile from '../utils/writeFile';
+// import { pathInSrc } from '../utils/default-path'
+import reactComp from '../templates/react/comp.js';
+import reactFuncComp from '../templates/react/func.js';
+import reactTest from '../templates/react/test.js';
+import writeFile from '../utils/write-file';
 
-const TEMPLATE_PATH = '../templates';
+
+// const TEMPLATE_PATH = pathInSrc('templates');
 
 const getFileExtension = ({ vue }: { vue: boolean }) => {
   return vue
@@ -9,22 +14,40 @@ const getFileExtension = ({ vue }: { vue: boolean }) => {
     : 'js';
 };
 
-const getTemplate = (templatePath: string, framework: string, type: string, append: string = '-comp') => {
-  return require(`${templatePath}/${framework}/${type}${append}`).default;
+const getTemplate = (framework: string, type: string) => {
+  if (framework === 'react') {
+    return type === 'comp'
+      ? reactComp
+      : reactComp === 'test'
+        ? reactFuncComp
+        : reactTest;
+  }
+
+  throw new Error('could not find path');
 };
 
+// const getTemplate = (templatePath: string, framework: string, type: string, append: string = '') => {
+//   return require(`${templatePath}/${framework}/${type}${append}`).default;
+// };
+
 export default async (framework: string, options: any) => {
-  const { name, test, dest } = options;
+  const { name, test, dest, functional } = options;
   if (!dest || !name) {
     console.log('Could not create amazing stuff due to some arguments missing: either name or dest');
     return;
   }
 
-  const fileTemplate = Mustache.render(getTemplate(TEMPLATE_PATH, framework, 'class'), options);
+  const type = functional
+    ? 'func'
+    : 'comp';
+
+  // const fileTemplate = Mustache.render(getTemplate(TEMPLATE_PATH, framework, type), options);
+  const fileTemplate = Mustache.render(getTemplate(framework, type), options);
   writeFile(fileTemplate, name, getFileExtension(options), dest);
 
   if (test) {
-    const testTemplate = Mustache.render(getTemplate(TEMPLATE_PATH, framework, 'test'), options);
+    // const testTemplate = Mustache.render(getTemplate(TEMPLATE_PATH, framework, 'test'), options);
+    const testTemplate = Mustache.render(getTemplate(framework, 'test'), options);
     writeFile(testTemplate, name, 'test.js', dest);
   }
